Ensure that brightness value is honored at turn on
Using "homeassistant.turn_on" does not support brightness. I changed it to `light.turn_on` to add support for brightness. I'm not sure if this breaks what this was initially intended for. I feel like the name "linked entities" is  more generic and adding support for brightness to it wasn't really appropriate. Perhaps, the proper name for a separate version of this is "linked lights."